Refactored(with EXCEPTION).
