+++ app/src/main/java/org/ecomap/android/app/fragments/EcoMapFragment.java
@@ -6,11 +6,7 @@ import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.database.Cursor;
-import android.graphics.Bitmap;
 import android.graphics.Color;
-import android.net.ConnectivityManager;
-import android.net.NetworkInfo;
-import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.design.widget.CoordinatorLayout;
 import android.support.design.widget.FloatingActionButton;
@@ -18,7 +14,6 @@ import android.support.design.widget.Snackbar;
 import android.support.design.widget.TabLayout;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
 import android.support.v4.content.LocalBroadcastManager;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AlertDialog;
@@ -27,11 +22,8 @@ import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.BaseAdapter;
-import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
-import android.widget.ProgressBar;
 import android.widget.ScrollView;
 import android.widget.TextView;
 import android.widget.Toast;
@@ -41,67 +33,43 @@ import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.MapView;
 import com.google.android.gms.maps.MapsInitializer;
 import com.google.android.gms.maps.UiSettings;
-import com.google.android.gms.maps.model.BitmapDescriptorFactory;
 import com.google.android.gms.maps.model.CameraPosition;
-import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.Marker;
-import com.google.android.gms.maps.model.MarkerOptions;
-import com.google.android.gms.maps.model.PolygonOptions;
-import com.google.maps.android.clustering.Cluster;
 import com.google.maps.android.clustering.ClusterManager;
-import com.nostra13.universalimageloader.core.DisplayImageOptions;
-import com.nostra13.universalimageloader.core.ImageLoader;
-import com.nostra13.universalimageloader.core.assist.FailReason;
-import com.nostra13.universalimageloader.core.listener.ImageLoadingProgressListener;
-import com.nostra13.universalimageloader.core.listener.SimpleImageLoadingListener;
 import com.wunderlist.slidinglayer.SlidingLayer;
 
-import org.ecomap.android.app.MyIconRendered;
 import org.ecomap.android.app.Problem;
 import org.ecomap.android.app.R;
 import org.ecomap.android.app.activities.MainActivity;
-import org.ecomap.android.app.activities.ViewPhotosActivity;
 import org.ecomap.android.app.data.EcoMapContract;
 import org.ecomap.android.app.data.model.ProblemPhotoEntry;
-import org.ecomap.android.app.sync.EcoMapAPIContract;
+import org.ecomap.android.app.sync.AddVoteTask;
 import org.ecomap.android.app.sync.EcoMapService;
+import org.ecomap.android.app.sync.GetPhotosTask;
+import org.ecomap.android.app.utils.ImageAdapter;
+import org.ecomap.android.app.utils.MapClustering;
+import org.ecomap.android.app.utils.NetworkAvailability;
+import org.ecomap.android.app.utils.PagerAdapter;
 import org.ecomap.android.app.widget.ExpandableHeightGridView;
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
 
 /**
  * A placeholder fragment containing a simple view.
  */
 public class EcoMapFragment extends Fragment {
-
-    public static final String ARG_SECTION_NUMBER = "ARG_SECTION_NUMBER";
     private static final String tag = "EcoMapFragment";
     private static int markerClickType;
     private static String filterCondition;
     public ImageAdapter imgAdapter;
-    Context mContext;
+    private Context mContext;
     Cursor cursor;
     EcoMapReceiver receiver;
-    Button cancelButton;
     MapView mapView;
-    private GoogleMap mMap; // Might be null if Google Play services APK is not available.
-    // Might be null if Google Play services APK is not available.
+    private GoogleMap mMap;
     private UiSettings UISettings;
     private ClusterManager<Problem> mClusterManager;
     private ArrayList<Problem> values;
-    private ArrayList<LatLng> points;
-    private ArrayList<Marker> markers;
     private View v;
     public SlidingLayer mSlidingLayer;
     private ImageView showTypeImage, showLike;
@@ -109,8 +77,7 @@ public class EcoMapFragment extends Fragment {
     private ScrollView detailedScrollView;
     private LinearLayout showHead;
     private Marker marker;
-    private List<String> mImagesURLArray;
-    private static CameraPosition cameraPosition;
+    public static CameraPosition cameraPosition;
 
     private FloatingActionButton fabAddProblem;
     private FloatingActionButton fabConfirm;
@@ -121,25 +88,15 @@ public class EcoMapFragment extends Fragment {
 
     //for rotating screen - save last position of SlidingPanel
     private static boolean isOpenSlidingLayer = false;
-    private static Problem lastOpenProblem;
+    public static Problem lastOpenProblem;
 
     CoordinatorLayout rootLayout;
     TabLayout tabLayout;
-
-    public static void setFilterCondition(String s) {
-        filterCondition = s;
-        Log.i(tag, "condition is set:" + filterCondition);
-    }
-
-    public static void setMarkerClickType(int type) {
-        markerClickType = type;
-    }
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         Log.i(tag, "onCreate");
-
     }
 
     @Override
@@ -147,12 +104,11 @@ public class EcoMapFragment extends Fragment {
         setRetainInstance(true);
         Log.i(tag, "onCreateView");
 
-
         v = inflater.inflate(R.layout.map_layout_main, container, false);
         mapView = (MapView) v.findViewById(R.id.mapview);
 
         //Temporary is to initialize mapView by null to get rotation works without exceptions.
-        mapView.onCreate(null);
+        //mapView.onCreate(null);
         Log.i(tag, "OnCreateView mapview");
 
         mMap = mapView.getMap();
@@ -167,8 +123,6 @@ public class EcoMapFragment extends Fragment {
         MapsInitializer.initialize(this.getActivity());
 
         values = new ArrayList<>();
-        points = new ArrayList<>();
-        markers = new ArrayList<>();
         mContext = getActivity();
 
         fabAddProblem = (FloatingActionButton) v.findViewById(R.id.fabAddProblem);
@@ -256,7 +210,6 @@ public class EcoMapFragment extends Fragment {
         mSlidingLayer.setOnInteractListener(new SlidingLayer.OnInteractListener() {
             @Override
             public void onOpen() {
-
                 //If onOpen, we show all lines of title
                 showTitle.setMaxLines(99);
                 showTitle.setEllipsize(null);
@@ -269,27 +222,21 @@ public class EcoMapFragment extends Fragment {
 
             @Override
             public void onShowPreview() {
-
                 fabAddProblem.setVisibility(View.INVISIBLE);
 
                 //If onPreview, we show only 1 line of title
                 showTitle.setMaxLines(1);
                 showTitle.setEllipsize(TextUtils.TruncateAt.END);
-
             }
 
             @Override
             public void onClose() {
-
                 fabAddProblem.setVisibility(View.VISIBLE);
-
             }
 
             @Override
             public void onOpened() {
-
                 isOpenSlidingLayer = true;
-
             }
 
             @Override
@@ -298,9 +245,7 @@ public class EcoMapFragment extends Fragment {
 
             @Override
             public void onClosed() {
-
                 isOpenSlidingLayer = false;
-
             }
         });
 
@@ -365,55 +310,36 @@ public class EcoMapFragment extends Fragment {
         mapView.onLowMemory();
     }
 
+    private class EcoMapReceiver extends BroadcastReceiver {
+        @Override
+        public void onReceive(Context context, Intent intent) {
+            fillMap(filterCondition);
+        }
+    }
+
     private void setUpMap() {
         Log.i(tag, "set up map");
 
-        if (isNetworkAvailable()) {
+        if (new NetworkAvailability(getActivity().getSystemService(Context.CONNECTIVITY_SERVICE))
+                .isNetworkAvailable()) {
             //Start service to get a new number of revision and new data
             Intent intent = new Intent(this.getActivity(), EcoMapService.class);
             getActivity().startService(intent);
         }
         else{
-            if (filterCondition == null) {
-                fillMap();
-            } else {
-                fillMap(filterCondition);
-            }
+            fillMap(filterCondition);
         }
-    }
-
-    public void fillMap() {
-        Log.i(tag, "fillmap");
-        double latitude, longitude;
-        String title;
-        int type_id;
-
-        values.clear();
-        mMap.clear();
-
-        cursor = getActivity().getContentResolver()
-                .query(EcoMapContract.ProblemsEntry.CONTENT_URI, null, null, null, null);
-
-        while (cursor.moveToNext()) {
-            Problem p = new Problem(cursor, getActivity());
-            values.add(p);
-        }
-
-        cursor.close();
-        setUpClusterer();
     }
 
     //if there is some filter condition, then this method will be called. It will search only needed points
     public void fillMap(String filterCondition) {
+        if (filterCondition == null){
+            filterCondition = "";
+        }
         Log.i(tag, "fillmap with condition" + filterCondition);
-        double latitude, longitude;
-        String title;
-        int type_id;
-
 
         values.clear();
         mMap.clear();
-
 
         cursor = getActivity().getContentResolver()
                 .query(EcoMapContract.ProblemsEntry.CONTENT_URI, null, filterCondition, null, null);
@@ -425,105 +351,25 @@ public class EcoMapFragment extends Fragment {
         }
 
         cursor.close();
-        setUpClusterer();
+        new MapClustering(cameraPosition, mMap, mContext, values, marker, this).setUpClusterer();
     }
 
-    public void setUpClusterer() {
 
-            if(cameraPosition != null){
 
-                mMap.moveCamera(CameraUpdateFactory.newCameraPosition(cameraPosition));
-            }
-
-            else {
-                //Position the map from static variables
-                double longitude = 30.417397;
-                double latitude = 50.461166;
-                float zoomlevel = 5;
-                mMap.moveCamera(CameraUpdateFactory.newLatLngZoom(new LatLng(latitude, longitude), zoomlevel));
-            }
-
-        //Initialize the manager with the mContext and the map.
-        mClusterManager = new ClusterManager<>(mContext, mMap);
-
-        //Point the map's listeners at the listeners implemented by the cluster.
-        mMap.setOnCameraChangeListener(mClusterManager);
-        mMap.setOnMarkerClickListener(mClusterManager);
-        mMap.setMyLocationEnabled(true);
-        mMap.setOnMarkerClickListener(mClusterManager);
-        mMap.setOnMapClickListener(new GoogleMap.OnMapClickListener() {
-            @Override
-            public void onMapClick(LatLng latLng) {
-                if (markerClickType == 1) {
-                    Marker m = mMap.addMarker(new MarkerOptions().position(latLng));
-                    m.setIcon(BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_ORANGE));
-                    markers.add(m);
-                    points.add(latLng);
-                    mMap.setOnMarkerClickListener(new GoogleMap.OnMarkerClickListener() {
-                        @Override
-                        public boolean onMarkerClick(Marker marker) {
-                            if (markerClickType == 1) {
-                                if (points.get(0).equals(marker.getPosition())) {
-                                    countPolygonPoints();
-                                    for (Marker m : markers) {
-                                        m.remove();
-                                    }
-                                }
-                                return false;
-                            }
-                            return false;
-                        }
-                    });
-
-                    //ADDING PROBLEM VIA FLOATING ACTION BUTTON
-
-                } else if (markerClickType == 2) {
-
-                    if (marker != null) {
-                        marker.remove();
-                    }
-
-                    marker = mMap.addMarker(new MarkerOptions().position(latLng));
-                    marker.setTitle("Houston we have a problem here!");
-                    marker.setIcon(BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_ORANGE));
-
-                }
-            }
-        });
-
-        //Add cluster items to the cluster manager.
-        mClusterManager.addItems(values);
-        mClusterManager.setOnClusterClickListener
-                (new ClusterManager.OnClusterClickListener<Problem>() {
-                    @Override
-                    public boolean onClusterClick(Cluster cluster) {
-                        mMap.moveCamera(CameraUpdateFactory.newLatLngZoom(cluster.getPosition(),
-                                mMap.getCameraPosition().zoom + 2));
-                        return false;
-                    }
-                });
-
-        //Set render for markers rendering
-        mClusterManager.setRenderer(new MyIconRendered(mContext, mMap, mClusterManager));
-
-        mClusterManager.setOnClusterItemClickListener(new ClusterManager.OnClusterItemClickListener<Problem>() {
-            @Override
-            public boolean onClusterItemClick(final Problem problem) {
-
-                fillSlidingPanel(problem);
-
-                //Set part of sliding layer visible
-                mSlidingLayer.openPreview(true);
-
-                //save last open Problem for rotating screen
-                lastOpenProblem = problem;
-
-                return false;
-            }
-        });
+    public static void setFilterCondition(String s) {
+        filterCondition = s;
+        Log.i(tag, "condition is set:" + filterCondition);
     }
 
-    private void fillSlidingPanel(final Problem problem){
+    public static void setMarkerClickType(int type) {
+        markerClickType = type;
+    }
+
+    public static int getMarkerClickType(){
+        return markerClickType;
+    }
+
+    public void fillSlidingPanel(final Problem problem){
 
         mMap.moveCamera(CameraUpdateFactory.newLatLngZoom(problem.getPosition(),
                /*mMap.getCameraPosition().zoom*/15.0f));
@@ -554,7 +400,7 @@ public class EcoMapFragment extends Fragment {
                     problem.setNumberOfLikes(1);
                     problem.setLiked(true);
 
-                    new AsyncAddVote().execute(problem.getId());
+                    new AddVoteTask().execute(problem.getId());
 
                     Toast.makeText(mContext, mContext.getString(R.string.message_isLiked), Toast.LENGTH_SHORT).show();
 
@@ -576,7 +422,7 @@ public class EcoMapFragment extends Fragment {
         chFm.beginTransaction().replace(R.id.fragment_comments, f, CommentsFragment.TAG).commit();
 
         //photos
-        new AsyncGetPhotos().execute(problem.getId());
+        new GetPhotosTask(this).execute(problem.getId());
     }
 
     private void signInAlertDialog() {
@@ -599,7 +445,7 @@ public class EcoMapFragment extends Fragment {
         alert.show();
     }
 
-    private void showTabLayout(){
+    /*private void showTabLayout(){
 
         if (tabLayout != null){
             tabLayout.removeAllTabs();
@@ -638,389 +484,6 @@ public class EcoMapFragment extends Fragment {
             }
         });
 
-    }
-
-    private void countPolygonPoints() {
-        if (points.size() >= 3) {
-            PolygonOptions polygonOptions = new PolygonOptions();
-            polygonOptions.addAll(points);
-            polygonOptions.strokeColor(mContext.getResources().getColor(R.color.polygonEdges));
-            polygonOptions.strokeWidth(7);
-            polygonOptions.fillColor(mContext.getResources().getColor(R.color.polygon));
-            mMap.addPolygon(polygonOptions);
-            points.clear();
-            markerClickType = 0;
-            mMap.setOnMarkerClickListener(mClusterManager);
-        }
-    }
-
-    private boolean isNetworkAvailable() {
-        ConnectivityManager connectivityManager
-                = (ConnectivityManager) getActivity().getSystemService(Context.CONNECTIVITY_SERVICE);
-        NetworkInfo activeNetworkInfo = connectivityManager.getActiveNetworkInfo();
-        return activeNetworkInfo != null && activeNetworkInfo.isConnected();
-    }
-
-    private static class ImageAdapter extends BaseAdapter {
-
-        private final Context mContext;
-        private List<ProblemPhotoEntry> mImagesURLArray;
-        private LayoutInflater inflater;
-        private DisplayImageOptions options;
-
-
-        public ImageAdapter(Context c, List<ProblemPhotoEntry> titledPhotos) {
-            this.mImagesURLArray = titledPhotos;
-            this.mContext = c;
-
-            this.inflater = LayoutInflater.from(mContext);
-
-            this.options = new DisplayImageOptions.Builder()
-                    //.showImageOnLoading(R.drawable.ic_stub)
-                    .showImageForEmptyUri(R.drawable.ic_empty)
-                            //.showImageOnFail(R.drawable.ic_action_refresh)
-                    .cacheInMemory(true)
-                    .cacheOnDisk(true)
-                    .considerExifParams(true)
-                    .bitmapConfig(Bitmap.Config.RGB_565)
-                    .build();
-        }
-
-        /**
-         * Update adapter data set
-         */
-        public void updateDataSet(List<ProblemPhotoEntry> data) {
-            mImagesURLArray = data;
-            notifyDataSetChanged();
-        }
-
-        @Override
-        public int getCount() {
-            return mImagesURLArray.size();
-        }
-
-        @Override
-        public Object getItem(int position) {
-            return null;
-        }
-
-        @Override
-        public long getItemId(int position) {
-            return position;
-        }
-
-        // create a new ImageView for each item referenced by the Adapter
-        @Override
-        public View getView(final int position, View convertView, ViewGroup parent) {
-
-            final ViewHolder holder;
-            View view = convertView;
-
-            if (view == null) {
-
-                view = inflater.inflate(R.layout.item_image_grid, parent, false);
-                holder = new ViewHolder();
-
-                assert view != null;
-
-                holder.imageView = (ImageView) view.findViewById(R.id.image);
-                holder.progressBar = (ProgressBar) view.findViewById(R.id.progress);
-
-                view.setTag(holder);
-
-            } else {
-                holder = (ViewHolder) view.getTag();
-            }
-
-            if (mImagesURLArray != null && mImagesURLArray.size() > 0) {
-                final ProblemPhotoEntry problemPhotoEntry = mImagesURLArray.get(position);
-
-                holder.imageView.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        Intent intent = new Intent(mContext, ViewPhotosActivity.class);
-                        intent.putExtra(ViewPhotosActivity.IMAGE_POSITION, position);
-                        intent.putExtra(ViewPhotosActivity.PHOTO_ENTRY, mImagesURLArray.toArray(new ProblemPhotoEntry[mImagesURLArray.size()]));
-                        mContext.startActivity(intent);
-                    }
-                });
-
-
-                /* On case they will change naming logic again
-                String[] imgName = problemPhotoEntry.getImgURL().split("\\.");
-                final String imgURL = EcoMapAPIContract.ECOMAP_HTTP_BASE_URL + "/static/thumbnails/" + imgName[0] + "." + "thumbnail." + imgName[1];
-                */
-
-                final String imgURL = EcoMapAPIContract.ECOMAP_HTTP_BASE_URL + "/static/thumbnails/" + problemPhotoEntry.getImgURL();
-
-                ImageLoader imageLoader = ImageLoader.getInstance();
-                imageLoader.displayImage(
-                        imgURL,
-                        holder.imageView,
-                        options,
-                        new SimpleImageLoadingListener() {
-                            @Override
-                            public void onLoadingStarted(String imageUri, View view) {
-                                holder.progressBar.setProgress(0);
-                                holder.progressBar.setVisibility(View.VISIBLE);
-                            }
-
-                            @Override
-                            public void onLoadingFailed(String imageUri, View view, FailReason failReason) {
-                                holder.progressBar.setVisibility(View.GONE);
-                            }
-
-                            @Override
-                            public void onLoadingComplete(String imageUri, View view, Bitmap loadedImage) {
-                                holder.progressBar.setVisibility(View.GONE);
-                            }
-                        },
-                        new ImageLoadingProgressListener() {
-                            @Override
-                            public void onProgressUpdate(String imageUri, View view, int current, int total) {
-                                holder.progressBar.setProgress(Math.round(100.0f * current / total));
-                            }
-                        });
-            }
-            return view;
-        }
-
-        static class ViewHolder {
-            ImageView imageView;
-            ProgressBar progressBar;
-        }
-
-    }
-
-    private class EcoMapReceiver extends BroadcastReceiver {
-        @Override
-        public void onReceive(Context context, Intent intent) {
-
-            if (filterCondition == null) {
-                fillMap();
-            } else {
-                fillMap(filterCondition);
-            }
-
-        }
-    }
-
-    private class AsyncGetPhotos extends AsyncTask<Integer, Integer, List<ProblemPhotoEntry>> {
-
-        private final String LOG_TAG = AsyncGetPhotos.class.getSimpleName();
-
-        String JSONStr = null;
-
-        @Override
-        protected List<ProblemPhotoEntry> doInBackground(Integer... params) {
-
-            Integer numProblem = params[0];
-
-            String ECOMAP_PHOTOS_URL = EcoMapAPIContract.ECOMAP_HTTP_BASE_URL + "/api/problems/" + numProblem + "/photos";
-            HttpURLConnection urlConnection = null;
-            BufferedReader reader = null;
-            List<ProblemPhotoEntry> ret = new ArrayList<>();
-
-            try {
-                // Getting input stream from URL
-
-                URL url = new URL(ECOMAP_PHOTOS_URL);
-                urlConnection = (HttpURLConnection) url.openConnection();
-                urlConnection.setRequestMethod("GET");
-                urlConnection.connect();
-
-                // Read the input stream into a String
-                InputStream inputStream = urlConnection.getInputStream();
-                StringBuilder buffer = new StringBuilder();
-                if (inputStream == null) {
-                    return ret;
-                }
-                reader = new BufferedReader(new InputStreamReader(inputStream));
-
-                String line;
-                while ((line = reader.readLine()) != null) {
-                    buffer.append(line + "\n");
-                }
-
-                if (buffer.length() == 0) {
-                    return ret;
-                }
-
-                JSONStr = buffer.toString();
-
-                // Starting method for parsing data from JSON and writing them to database
-                ret = getPhotosFromJSON();
-
-            } catch (IOException e) {
-                Log.e(LOG_TAG, e.getMessage(), e);
-            } finally {
-                if (urlConnection != null) {
-                    urlConnection.disconnect();
-                }
-                if (reader != null) {
-                    try {
-                        reader.close();
-                    } catch (IOException e) {
-                        Log.e(LOG_TAG, e.getMessage(), e);
-                    }
-                }
-            }
-
-            return ret;
-        }
-
-        // parsing data from JSON and writing them to database
-        private List<ProblemPhotoEntry> getPhotosFromJSON() {
-            final String COMMENT = "comment";
-            final String PHOTO_NAME = "name";
-
-            try {
-                JSONArray jArr = new JSONArray(JSONStr);
-
-                List<ProblemPhotoEntry> syncedList = Collections.synchronizedList(new ArrayList<ProblemPhotoEntry>(JSONStr.length()));
-
-                for (int i = 0; i < jArr.length(); i++) {
-                    String title;
-                    String image_name;
-
-                    JSONObject obj = jArr.getJSONObject(i);
-                    title = obj.getString(COMMENT);
-                    image_name = obj.getString(PHOTO_NAME);
-                    syncedList.add(new ProblemPhotoEntry(title, image_name));
-                }
-
-                return syncedList;
-
-            } catch (JSONException e) {
-                e.printStackTrace();
-            }
-
-            return new ArrayList<ProblemPhotoEntry>();
-        }
-
-
-        @Override
-        protected void onPostExecute(List<ProblemPhotoEntry> imgagesArray) {
-            imgAdapter.updateDataSet(imgagesArray);
-        }
-    }
-
-    private class AsyncAddVote extends AsyncTask<Integer, Void, Boolean> {
-
-        private final String LOG_TAG = AsyncAddVote.class.getSimpleName();
-        private Integer problem_id;
-
-        @Override
-        protected Boolean doInBackground(Integer... params) {
-            URL url = null;
-            Boolean result = Boolean.FALSE;
-            problem_id = params[0];
-
-            //validation
-            if (MainActivity.isUserIsAuthorized()) {
-                if (params.length > 0 && params[0] != null) {
-
-                    HttpURLConnection connection = null;
-
-                    try {
-
-                        //creating JSONObject for request
-                        JSONObject request = new JSONObject();
-                        request.put("content", params[0]);
-
-                        url = new URL(EcoMapAPIContract.ECOMAP_API_URL + "/problems/" + problem_id + "/vote");
-
-                        connection = (HttpURLConnection) url.openConnection();
-                        connection.setRequestMethod("POST");
-                        connection.setRequestProperty("Content-Type", "application/json; charset=UTF-8");
-                        connection.setDoOutput(true);
-                        connection.connect();
-
-                        /**
-                         * sending request
-                         * request.toString() - translate our object into appropriate JSON text
-                         * {
-                         *      "content": "your comment"
-                         * }
-                         */
-                        //OutputStream outputStream = connection.getOutputStream();
-                        //outputStream.write(request.toString().getBytes("UTF-8"));
-
-                        //handling result from server
-                        if (connection.getResponseCode() == HttpURLConnection.HTTP_OK) {
-                            result = Boolean.TRUE;
-                        }
-
-                    } catch (Exception e) {
-                        Log.e(LOG_TAG, e.getMessage(), e);
-                    } finally {
-                        if (connection != null) {
-                            connection.disconnect();
-                        }
-                    }
-                }
-            }
-
-            return result;
-        }
-
-
-    }
-
+    }*/
 }
 
-    class PagerAdapter extends FragmentPagerAdapter {
-
-        int numOfTabs;
-        FragmentManager fragmentManager;
-
-        public PagerAdapter (FragmentManager fm, int numOfTabs){
-            super(fm);
-            this.numOfTabs = numOfTabs;
-            this.fragmentManager = fm;
-        }
-
-        @Override
-        public Fragment getItem(int index) {
-            Fragment fragment = null;
-            String tag = null;
-            switch (index){
-                case 0:
-    /*
-                        tag = EcoMapFragment.class.getSimpleName();
-                        fragment = fragmentManager.findFragmentByTag(tag);
-                        if (fragment == null) {
-                            fragment = new EcoMapFragment();
-                            return fragment;
-                        } else {
-                            return fragment;
-                        }
-    */
-
-                    return new EcoMapFragment();
-
-                case 1:
-                    tag = AddProblemFragment.class.getSimpleName();
-                    fragment = fragmentManager.findFragmentByTag(tag);
-                    if (fragment == null) {
-                        fragment = new AddProblemFragment();
-                    }
-
-                    return fragment;
-            }
-    /*
-                Bundle args = new Bundle();
-                args.putDoubleArray("Position", EcoMapFragment.getMarkerPosition());
-                fragment.setArguments(args);
-                android.support.v4.app.FragmentTransaction transaction = fragmentManager.beginTransaction();
-                transaction.addToBackStack(null);
-                transaction.replace(R.id.content_frame, fragment, tag).commit();
-    */
-            return null;
-        }
-
-        @Override
-        public int getCount() {
-            return 2;
-        }
-    }
